+++ client/src/lib/state/cos.js
@@ -20,6 +20,7 @@ const {
   titleCaseRender,
   kebabCaseInput,
   unconditionalInvalidText,
+  hideWhenUseData,
 } = require("./utils");
 const { cosPlans } = require("../constants");
 const { nameField, hideWhenFieldFalse } = require("./reusable-fields");
@@ -336,7 +337,9 @@ function initObjectStorageStore(store) {
               }`;
         },
       }),
-      resource_group: resourceGroupsField(),
+      resource_group: resourceGroupsField(false, {
+        noHideWhen: true,
+      }),
       kms: {
         type: "select",
         default: "",
@@ -372,6 +375,7 @@ function initObjectStorageStore(store) {
         onInputChange: kebabCaseInput("plan"),
         invalid: fieldIsNullOrEmptyString("plan"),
         invalidText: selectInvalidText("plan"),
+        hideWhen: hideWhenUseData,
       },
     },
     subComponents: {